src/requirements.txt
django>=1.8,<2
django-autocomplete-light>=3.2.10
django-admin-list-filter-dropdown>=1.0.1
django-admin-tools>=0.8.1
django-debug-toolbar
google-auth
google-auth-httplib2
google-api-python-client
social-auth-app-django
psycopg2-binary
gunicorn
django-filer